# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StackClass API Reference",
    "description": "Backend API and services for StackClass",
    "license": {
      "name": ""
    },
    "version": "0.19.0"
  },
  "paths": {
...
    "/v1/courses/{slug}/stages": {
      "get": {
        "tags": [
          "Stage"
        ],
        "summary": "Find all stages for a course (including extensions)",
        "operationId": "find-all-stages",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StageResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to get course"
          }
        }
      }
    },
    "/v1/courses/{slug}/stages/base": {
      "get": {
        "tags": [
          "Stage"
        ],
        "summary": "Find only base stages for a course (excluding extensions).",
        "operationId": "find-base-stages",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StageResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to get course"
          }
        }
      }
    },
    "/v1/courses/{slug}/stages/extended": {
      "get": {
        "tags": [
          "Stage"
        ],
        "summary": "Find only extended stages for a course.",
        "operationId": "find-extended-stages",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StageResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to get course"
          }
        }
      }
    },
    "/v1/courses/{slug}/stages/{stage_slug}": {
      "get": {
        "tags": [
          "Stage"
        ],
        "summary": "Get the details of the stage.",
        "operationId": "get-stage-detail",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage_slug",
            "in": "path",
            "description": "The slug of stage",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stage retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course or stage not found"
          },
          "500": {
            "description": "Failed to get course or stage"
          }
        }
      }
    },
...
    "/v1/user/courses/{slug}/stages": {
      "get": {
        "tags": [
          "User",
          "Stage"
        ],
        "summary": "Find all stages for the current user.",
        "operationId": "find-user-stages",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserStageResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to get course"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "User",
          "Stage"
        ],
        "summary": "Mark a stage as completed for the current user.",
        "operationId": "complete-stage",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Complete stage request",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "slug"
                ],
                "properties": {
                  "slug": {
                    "type": "string",
                    "description": "The slug of the stage to mark as completed"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stage completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course or stage not found"
          },
          "500": {
            "description": "Failed to complete stage"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      }
    },
    "/v1/user/courses/{slug}/stages/{stage_slug}": {
      "get": {
        "tags": [
          "User",
          "Stage"
        ],
        "summary": "Get the details of the stage for the current user.",
        "operationId": "get-user-stage-detail",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage_slug",
            "in": "path",
            "description": "The slug of stage",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stage retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course or stage not found"
          },
          "500": {
            "description": "Failed to get course or stage"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "StageDetailResponse": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "difficulty",
          "description",
          "instruction",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "description": {
            "type": "string",
            "description": "A short markdown description of the stage,\nused in the course overview page."
          },
          "difficulty": {
            "type": "string",
            "description": "Difficulty level (very_easy, easy, medium, hard)"
          },
          "extension_slug": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional slug of the parent extension (null if part of main course)"
          },
          "instruction": {
            "type": "string",
            "description": "A markdown description for this stage."
          },
          "name": {
            "type": "string",
            "description": "Display name of the stage"
          },
          "slug": {
            "type": "string",
            "description": "Unique human-readable identifier within parent context"
          },
          "solution": {
            "type": [
              "string",
              "null"
            ],
            "description": "The solution to this stage, if available."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "StageResponse": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "difficulty",
          "description",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "description": {
            "type": "string",
            "description": "A short markdown description of the stage,\nused in the course overview page."
          },
          "difficulty": {
            "type": "string",
            "description": "Difficulty level (very_easy, easy, medium, hard)"
          },
          "extension_slug": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional slug of the parent extension (null if part of main course)"
          },
          "name": {
            "type": "string",
            "description": "Display name of the stage"
          },
          "slug": {
            "type": "string",
            "description": "Unique human-readable identifier within parent context"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "UserStageResponse": {
        "type": "object",
        "required": [
          "course_slug",
          "stage_slug",
          "status",
          "started_at"
        ],
        "properties": {
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Timestamp when the stage was completed"
          },
          "course_slug": {
            "type": "string",
            "description": "Slug of the enrolled course"
          },
          "stage_slug": {
            "type": "string",
            "description": "Slug of the stage"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the stage was started"
          },
          "status": {
            "type": "string",
            "description": "Current progress status (in_progress, completed)"
          }
        }
      }
    },
    "securitySchemes": {
      "JWTBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}